requirements.txt
Django>=2.0.13
django-cors-headers>=3.0.2
djangorestframework>=3.9.4
psycopg2>=2.8.3
requests==2.22.0
